+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### New features
 
+* [#490](https://github.com/clojure-emacs/cider/pull/490) Dedicated
+  support for `company-mode` in `cider-complete-at-point`.
 * [#460](https://github.com/clojure-emacs/cider/issues/460) Support for
 cider-nrepl's complete middleware for CLJ/CLJS autocomplete.
 * [#465](https://github.com/clojure-emacs/cider/issues/465) Support for